pyproject.toml
[tool.poetry]
name = "cobrapy"
version = "0.1.0"
description = "Content Based Image Retrieval Analysis - a python package for parsing videos using multi-modal large langauge models"
authors = ["Brady Leavitt <brady.d.leavitt@gmail.com>","Owen Van Valkenburg <owenv@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.37.0"
moviepy = "^1.0.3"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
numpy = "^2.0.1"
nest-asyncio = "^1.6.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest = "^8.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


